refactor: Rename HSKHanzi class file to hsk_hanzi.py
